+++ js/detail-pdf.js
@@ -1709,6 +1709,7 @@ views.beheer = renderBeheer;
 views.clubbeheer = renderClubBeheer;
 views.clubsadmin = renderClubsAdmin;
 views.allusers = renderAllUsers;
+views.online = renderOnline;
 views.teams = renderTeamsList;
 views.teamEdit = renderTeamEdit;
 views.tournaments = renderTournamentList;